refactor: extract game types and helpers
