+++ app/game/useGameState.ts
@@ -1,0 +1,97 @@
+"use client"
+
+import { useState } from "react"
+import type { GameState } from "@/lib/game/types"
+import { WEEKLY_TEMPLATES } from "@/lib/game/helpers"
+
+const defaultState: GameState = {
+  childName: "태기",
+  age: 8,
+  month: 3,
+  year: 1,
+  region: "seoul",
+  level: 1,
+  experience: 0,
+  skillPoints: 3,
+  maxExperience: 100,
+  physical: 30,
+  technique: 20,
+  intelligence: 40,
+  social: 35,
+  creativity: 25,
+  discipline: 30,
+  confidence: 40,
+  leadership: 15,
+  shooting: 15,
+  passing: 20,
+  dribbling: 25,
+  defending: 10,
+  goalkeeping: 5,
+  speed: 30,
+  stamina: 25,
+  mentality: 20,
+  health: 80,
+  stress: 10,
+  happiness: 70,
+  fatigue: 20,
+  motivation: 60,
+  energy: 100,
+  parentRelation: 80,
+  friendRelation: 50,
+  coachRelation: 60,
+  teammateRelation: 45,
+  teacherRelation: 70,
+  allowance: 50000,
+  savings: 0,
+  schoolName: "서울중앙초",
+  schoolType: "elementary",
+  grade: "2학년",
+  schoolRank: 15,
+  academicGrade: "B+",
+  attendanceRate: 95,
+  currentTeam: "홍대 골목구장",
+  position: "자유포지션",
+  reputation: 10,
+  matchesPlayed: 0,
+  goals: 0,
+  assists: 0,
+  wins: 0,
+  losses: 0,
+  draws: 0,
+  scoutingOffers: [],
+  streetCredits: 50,
+  trophies: [],
+  achievements: [],
+  records: [],
+  memories: [],
+  growthHistory: [],
+  weeklyTemplate: WEEKLY_TEMPLATES.street_warrior.schedule,
+  monthlySchedule: {},
+  currentDay: 1,
+  isMonthRunning: false,
+  monthProgress: 0,
+  isInjured: false,
+  injuryType: undefined,
+  injuryName: undefined,
+  injuryDays: 0,
+  totalInjuryDays: 0,
+  specialEvents: [],
+  upcomingMatches: [],
+  careerPath: "undecided",
+  graduationYear: 0,
+  currentSeason: "spring",
+  personalityTraits: {
+    competitiveness: 60,
+    teamwork: 45,
+    creativity: 70,
+    discipline: 40,
+    confidence: 55,
+    calmness: 35,
+  },
+  rivals: [],
+}
+
+export function useGameState(initial?: GameState) {
+  const [state, setState] = useState<GameState>(initial ?? defaultState)
+  return { state, setState }
+}
